+++ src/tsa/regularization.cpp
@@ -8,7 +8,7 @@
 
 af::array tsa::regularization::groupBy(af::array in, af::array (*aggregationFunction)(const af::array &, const dim_t),
                                        int nColumnsKey, int nColumnsValue) {
-    long n = in.dims(0);
+    long n = static_cast<long>(in.dims(0));
 
     af::seq keyColumns = af::seq(nColumnsKey);
     af::seq valuesColumns = af::seq(nColumnsValue) + (double)nColumnsKey;
@@ -22,7 +22,7 @@ af::array tsa::regularization::groupBy(af::array in, af::array (*aggregationFunc
     // cannot be used inside
     // gfor(af::seq i, groupKeys.dims(0)) {
     for (int i = 0; i < groupKeys.dims(0); i++) {
-        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), in.dims(0));
+        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), static_cast<const unsigned int>(in.dims(0)));
         af::array valuesMask = af::allTrue(in(af::span, keyColumns) == tiledGroupKey, 1);
         values(i, af::span) += aggregationFunction(in(valuesMask, valuesColumns), 0);
     }
@@ -33,7 +33,7 @@ af::array tsa::regularization::groupBy(af::array in, af::array (*aggregationFunc
 af::array tsa::regularization::groupBy(af::array in,
                                        af::array (*aggregationFunction)(const af::array &, bool, const dim_t),
                                        int nColumnsKey, int nColumnsValue) {
-    long n = in.dims(0);
+    long n = static_cast<long>(in.dims(0));
 
     af::seq keyColumns = af::seq(nColumnsKey);
     af::seq valuesColumns = af::seq(nColumnsValue) + (double)nColumnsKey;
@@ -47,7 +47,7 @@ af::array tsa::regularization::groupBy(af::array in,
     // cannot be used inside
     // gfor(af::seq i, groupKeys.dims(0)) {
     for (int i = 0; i < groupKeys.dims(0); i++) {
-        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), in.dims(0));
+        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), static_cast<const unsigned int>(in.dims(0)));
         af::array valuesMask = af::allTrue(in(af::span, keyColumns) == tiledGroupKey, 1);
         values(i) += aggregationFunction(in(valuesMask, valuesColumns), true, 0);
     }
@@ -57,7 +57,7 @@ af::array tsa::regularization::groupBy(af::array in,
 
 af::array tsa::regularization::groupBy(af::array in, af::array (*aggregationFunction)(const af::array &, const int),
                                        int nColumnsKey, int nColumnsValue) {
-    long n = in.dims(0);
+    long n = static_cast<long>(in.dims(0));
 
     af::seq keyColumns = af::seq(nColumnsKey);
     af::seq valuesColumns = af::seq(nColumnsValue) + (double)nColumnsKey;
@@ -71,7 +71,7 @@ af::array tsa::regularization::groupBy(af::array in, af::array (*aggregationFunc
     // cannot be used inside
     // gfor(af::seq i, groupKeys.dims(0)) {
     for (int i = 0; i < groupKeys.dims(0); i++) {
-        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), in.dims(0));
+        af::array tiledGroupKey = af::tile(groupKeys(i, af::span), static_cast<const unsigned int>(in.dims(0)));
         af::array valuesMask = af::allTrue(in(af::span, keyColumns) == tiledGroupKey, 1);
         values(i) += aggregationFunction(in(valuesMask, valuesColumns), 0);
     }
