+++ src/tsa/statistics.cpp
@@ -8,7 +8,7 @@
 #include <tsa/statistics.h>
 
 af::array tsa::statistics::covariance(af::array tss, bool unbiased) {
-    long n = tss.dims(0);
+    long n = static_cast<long>(tss.dims(0));
 
     af::array result = tsa::features::crossCovariance(tss, tss) * n / (n - unbiased);
 
@@ -16,24 +16,25 @@ af::array tsa::statistics::covariance(af::array tss, bool unbiased) {
 }
 
 af::array tsa::statistics::moment(af::array tss, int k) {
-    double n = tss.dims(0);
+    double n = static_cast<double>(tss.dims(0));
 
     return af::sum(af::pow(tss, k), 0) / n;
 }
 
 af::array tsa::statistics::sampleStdev(af::array tss) {
-    double n = tss.dims(0);
+    double n = static_cast<double>(tss.dims(0));
     af::array mean = af::mean(tss, 0);
 
-    return af::sqrt(af::sum(af::pow(tss - af::tile(mean, tss.dims(0)), 2), 0) / (n - 1));
+    return af::sqrt(af::sum(af::pow(tss - af::tile(mean, static_cast<const unsigned int>(tss.dims(0))), 2), 0) /
+                    (n - 1));
 }
 
 af::array tsa::statistics::kurtosis(af::array tss) {
-    double n = tss.dims(0);
+    double n = static_cast<double>(tss.dims(0));
 
     double a = (n * (n + 1)) / ((n - 1) * (n - 2) * (n - 3));
-    af::array tssMinusMean = (tss - af::tile(af::mean(tss, 0), tss.dims(0)));
-    af::array sampleStdev = af::tile(tsa::statistics::sampleStdev(tss), tss.dims(0));
+    af::array tssMinusMean = (tss - af::tile(af::mean(tss, 0), static_cast<const unsigned int>(tss.dims(0))));
+    af::array sampleStdev = af::tile(tsa::statistics::sampleStdev(tss), static_cast<const unsigned int>(tss.dims(0)));
     af::array b = af::sum(af::pow(tssMinusMean / sampleStdev, 4), 0);
     double c = (3 * std::pow(n - 1, 2)) / ((n - 2) * (n - 3));
 
@@ -41,30 +42,31 @@ af::array tsa::statistics::kurtosis(af::array tss) {
 }
 
 af::array tsa::statistics::skewness(af::array tss) {
-    float n = tss.dims(0);
-    af::array tssMinusMean = (tss - af::tile(af::mean(tss, 0), tss.dims(0)));
+    float n = static_cast<float>(tss.dims(0));
+    af::array tssMinusMean = (tss - af::tile(af::mean(tss, 0), static_cast<const unsigned int>(tss.dims(0))));
     af::array m3 = tsa::statistics::moment(tssMinusMean, 3);
     af::array s3 = af::pow(tsa::statistics::sampleStdev(tss), 3);
     return (std::pow(n, 2.0) / ((n - 1) * (n - 2))) * m3 / s3;
 }
 
 af::array tsa::statistics::quantile(af::array tss, af::array q, float precision) {
-    long n = tss.dims(0);
+    long n = static_cast<long>(tss.dims(0));
 
     af::array idx = q * (n - 1);
     af::array idxAsInt = idx.as(af::dtype::u32);
 
-    af::array a = af::tile(idxAsInt == idx, 1, tss.dims(1)) * tss(idx, af::span);
+    af::array a = af::tile(idxAsInt == idx, 1, static_cast<const unsigned int>(tss.dims(1))) * tss(idx, af::span);
     af::array fraction = (idx * precision) % precision / precision;
 
-    af::array b = af::tile(idxAsInt != idx, 1, tss.dims(1)) * tss(idxAsInt, af::span) +
-                  (tss(idxAsInt + 1, af::span) - tss(idxAsInt, af::span)) * af::tile(fraction, 1, tss.dims(1));
+    af::array b = af::tile(idxAsInt != idx, 1, static_cast<const unsigned int>(tss.dims(1))) * tss(idxAsInt, af::span) +
+                  (tss(idxAsInt + 1, af::span) - tss(idxAsInt, af::span)) *
+                      af::tile(fraction, 1, static_cast<const unsigned int>(tss.dims(1)));
     return a + b;
 }
 
 af::array searchSorted(af::array tss, af::array qs) {
-    af::array input = af::tile(tss, 1, 1, qs.dims(0));
-    af::array qsReordered = af::tile(af::reorder(qs, 2, 1, 0, 3), tss.dims(0));
+    af::array input = af::tile(tss, 1, 1, static_cast<const unsigned int>(qs.dims(0)));
+    af::array qsReordered = af::tile(af::reorder(qs, 2, 1, 0, 3), static_cast<const unsigned int>(tss.dims(0)));
 
     af::array result = af::flip(af::sum(input < qsReordered, 2), 0);
 
@@ -81,8 +83,10 @@ af::array tsa::statistics::quantilesCut(af::array tss, float quantiles, float pr
     af::array ss = searchSorted(tss, qs);
 
     af::array qcut = af::array(qs.dims(0) - 1, 2, qs.dims(1), qs.type());
-    qcut(af::span, 0, af::span) = af::reorder(qs(af::seq(0, qs.dims(0) - 2), af::span), 0, 2, 1, 3);
-    qcut(af::span, 1, af::span) = af::reorder(qs(af::seq(1, qs.dims(0) - 1), af::span), 0, 2, 1, 3);
+    qcut(af::span, 0, af::span) =
+        af::reorder(qs(af::seq(0, static_cast<double>(qs.dims(0)) - 2), af::span), 0, 2, 1, 3);
+    qcut(af::span, 1, af::span) =
+        af::reorder(qs(af::seq(1, static_cast<double>(qs.dims(0)) - 1), af::span), 0, 2, 1, 3);
 
     qcut(0, 0, af::span) -= precision;
